requirements.txt
# Core ML/RL libraries
torch==2.0.1
stable-baselines3==2.1.0
gymnasium==0.29.1

# NLP and paper processing
transformers==4.35.0
sentence-transformers==2.2.2
arxiv==2.0.0
requests==2.31.0
beautifulsoup4==4.12.2

# Data processing
pandas==2.1.3
numpy==1.24.3
scipy==1.11.4
networkx==3.2

# Visualization
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0
tensorboard==2.15.0

# API (optional for now)
fastapi==0.104.1
uvicorn==0.24.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Utilities
tqdm==4.66.1
python-dotenv==1.0.0
pyyaml==6.0.1
